AI-Visual-Defect-Detection-AutoModel/requirements.txt
torch
torchvision
opencv-python
numpy
pandas
matplotlib
scikit-learn
streamlit
Pillow
